+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortMcpServerConfigTemplate.java
@@ -1,6 +1,0 @@
-package ai.mcpdirect.backend.dao.entity.aitool;
-
-public class AIPortMcpServerConfigTemplate {
-    public long id;
-    public String config;
-}