+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortToolMakerTemplate.java
@@ -9,4 +9,6 @@ public class AIPortToolMakerTemplate {
     public long agentId;
     public long lastUpdated;
     public long userId;
+    public String config;
+    public String inputs;
 }