+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortToolMakerTemplateInstance.java
@@ -1,0 +1,10 @@
+package ai.mcpdirect.backend.dao.entity.aitool;
+
+public class AIPortToolMakerTemplateInstance {
+    public long id;
+    public long templateId;
+    public String inputs;
+    public long userId;
+    public long lastUpdated;
+    public long created;
+}